Compiled handheld soil moisture
Here is the end result of cleaning and processing the handheld soil moisture data. This does not include the entire data set from 2018-08-17, part of which had to be thrown out because no standard error could be calculated. It also contains none of the data from 2018-11-01 because it only included data from part of shelter 1.